requirements-dev.txt
pytest
httpx
black
pytest-asyncio
python-dotenv
